fixes for plans by equipment specification, new specification for likes and saves

+++ src/main/java/source/code/repository/RecipeFoodRepository.java
@@ -6,7 +6,7 @@ import source.code.model.Recipe.RecipeFood;
 import java.util.List;
 import java.util.Optional;
 
-public interface RecipeFoodRepository extends JpaRepository<RecipeFood, Long> {
+public interface RecipeFoodRepository extends JpaRepository<RecipeFood, Integer> {
   List<RecipeFood> findByRecipeId(int recipeId);
   List<RecipeFood> findByFoodId(int foodId);
   boolean existsByRecipeIdAndFoodId(int recipeId, int foodId);
